+++ .gitignore
@@ -10,3 +10,5 @@ __pycache__
 main
 run.sh
 server
+dllama
+dllama-*
