+++ README.md
@@ -127,7 +127,7 @@ git clone https://github.com/b4rtaz/distributed-llama.git
 ```
 6. Compile Distributed Llama:
 ```sh
-make main
+make dllama
 ```
 7. Transfer weights and the tokenizer file to the root device.
 8. Optional: assign static IP addresses.
@@ -137,17 +137,17 @@ sudo ip addr add 10.0.0.2/24 dev eth0 # 2th device
 ```
 9. Run worker nodes on worker devices:
 ```sh
-sudo nice -n -20 ./main worker --port 9998 --nthreads 4
+sudo nice -n -20 ./dllama worker --port 9998 --nthreads 4
 ```
 10. Run root node on the root device:
 ```sh
-sudo nice -n -20 ./main inference --model ../dllama_llama-2-7b_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4 --workers 10.0.0.2:9998
+sudo nice -n -20 ./dllama inference --model ../dllama_llama-2-7b_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4 --workers 10.0.0.2:9998
 ```
 
 To add more worker nodes, just add more addresses to the `--workers` argument.
 
 ```
-./main inference ... --workers 10.0.0.2:9998 10.0.0.3:9998 10.0.0.4:9998
+./dllama inference ... --workers 10.0.0.2:9998 10.0.0.3:9998 10.0.0.4:9998
 ```
 
 [Share your results](https://github.com/b4rtaz/distributed-llama/discussions)!
@@ -166,20 +166,20 @@ git clone https://github.com/b4rtaz/distributed-llama.git
 ```
 3. Compile Distributed Llama:
 ```sh
-make main
+make dllama
 ```
 4. Transfer weights and the tokenizer file to the root node.
 5. Run worker nodes on worker devices:
 ```sh
-sudo nice -n -20 ./main worker --port 9998 --nthreads 4
+sudo nice -n -20 ./dllama worker --port 9998 --nthreads 4
 ```
 6. Run root node on the root device:
 ```sh
-sudo nice -n -20 ./main inference --model ../dllama_llama-2-7b_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4 --workers 192.168.0.1:9998
+sudo nice -n -20 ./dllama inference --model ../dllama_llama-2-7b_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4 --workers 192.168.0.1:9998
 ```
 7. To run the root node in the chat mode:
 ```sh
-sudo nice -n -20 ./main chat --model ../dllama_llama-2-7b-chat_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --nthreads 4 --workers 192.168.0.1:9998
+sudo nice -n -20 ./dllama chat --model ../dllama_llama-2-7b-chat_q40.bin --tokenizer ../dllama-llama2-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --nthreads 4 --workers 192.168.0.1:9998
 ```
 
 [Share your results](https://github.com/b4rtaz/distributed-llama/discussions)!
