+++ docs/GROK.md
@@ -11,9 +11,9 @@ wget https://huggingface.co/b4rtaz/grok-1-distributed-llama/resolve/main/dllama-
 ```
 4. Build the project:
 ```bash
-make main
+make dllama
 ```
 5. Run the model:
 ```bash
-./main inference --weights-float-type q40 --buffer-float-type q80 --prompt "Hello" --steps 128 --nthreads 8 --model dllama-grok-1-q40-final.bin --tokenizer dllama-grok1-tokenizer.t
+./dllama inference --weights-float-type q40 --buffer-float-type q80 --prompt "Hello" --steps 128 --nthreads 8 --model dllama-grok-1-q40-final.bin --tokenizer dllama-grok1-tokenizer.t
 ```
